# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Indexer APIs for Andamio dapp-system.",
        "title": "Andamio Indexer 1.0.0",
        "contact": {
            "name": "Andamio Support",
            "url": "https://www.andamio.com/support",
            "email": "dev@andamio.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "0.0.0.0:42069",
    "basePath": "/api/v1/indexer",
    "paths": {
...
        "/transactions/{tx_hash}/inputs/{input_index}/datum": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves the datum associated with a specific transaction input.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get Datum for Transaction Input",
                "operationId": "getInputDatum",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The transaction hash (hex-encoded).",
                        "name": "tx_hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The index of the transaction input.",
                        "name": "input_index",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved datum.",
                        "schema": {
                            "$ref": "#/definitions/viewmodel.Datum"
                        }
                    },
                    "400": {
                        "description": "Invalid transaction hash or input index.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Datum not found for the specified input.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transactions/{tx_hash}/inputs/{input_index}/redeemer": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves the redeemer associated with a specific transaction input.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get Redeemer for Transaction Input",
                "operationId": "getInputRedeemer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The transaction hash (hex-encoded).",
                        "name": "tx_hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The index of the transaction input.",
                        "name": "input_index",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved redeemer.",
                        "schema": {
                            "$ref": "#/definitions/viewmodel.Redeemer"
                        }
                    },
                    "400": {
                        "description": "Invalid transaction hash or input index.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Redeemer not found for the specified input.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transactions/{tx_hash}/outputs/{output_index}/datum": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves the datum associated with a specific transaction output.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get Datum for Transaction Output",
                "operationId": "getOutputDatum",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The transaction hash (hex-encoded).",
                        "name": "tx_hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The index of the transaction output.",
                        "name": "output_index",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved datum.",
                        "schema": {
                            "$ref": "#/definitions/viewmodel.Datum"
                        }
                    },
                    "400": {
                        "description": "Invalid transaction hash or output index.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Datum not found for the specified output.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "viewmodel.Datum": {
            "type": "object",
            "properties": {
                "datum_cbor": {
                    "type": "string"
                },
                "datum_hash": {
                    "type": "string"
                },
                "utxo_id": {
                    "type": "string"
                },
                "utxo_index": {
                    "type": "integer"
                }
            }
        },
        "viewmodel.Redeemer": {
            "type": "object",
            "properties": {
                "cbor": {
                    "description": "CBOR string representation",
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "tag": {
                    "type": "integer"
                },
                "transaction_hash": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}